requirement.txt
appnope==0.1.0
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.0
common==0.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
Django==1.11.6
django-cors-headers==3.1.1
django-extensions==1.9.8
django-qrcode==0.3
django-rest-swagger==2.2.0
djangorestframework==3.11.1
idna==2.7
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.13.1
Jinja2==2.11.2
MarkupSafe==1.1.1
olefile==0.44
openapi-codec==1.3.2
parso==0.3.1
pexpect==4.3.1
pickleshare==0.7.5
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.6.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.5
qrcode==5.3
requests==2.20.1
simplegeneric==0.8.1
simplejson==3.17.2
six==1.11.0
South==1.0.2
sqlparse==0.4.1
traitlets==4.3.2
typing==3.6.6
uritemplate==3.0.1
urllib3==1.22
wcwidth==0.1.7
